rest api v2

+++ myFirst_app/controllers/admin.js
@@ -1,9 +1,11 @@
 const moongose = require("mongoose");
+const fileHelper = require("../util/file");
 
 const Product = require("../models/product");
 const { validationResult } = require("express-validator");
 
 const { throwError } = require("../util/functions");
+const product = require("../models/product");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -22,7 +24,7 @@ exports.postAddProduct = (req, res, next) => {
   const price = req.body.price;
   const description = req.body.description;
 
-  if(!image) {
+  if (!image) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/add-product",
@@ -39,7 +41,7 @@ exports.postAddProduct = (req, res, next) => {
   }
   const imageUrl = image.path;
   const errors = validationResult(req);
-  
+
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
@@ -148,7 +150,8 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      if(updatedImage) {
+      if (updatedImage) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = updatedImage.path;
       }
       return product.save().then((response) => {
@@ -174,13 +177,20 @@ exports.getProducts = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.id;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then((response) => {
-      res.redirect("/admin/products");
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
+    .then(() => {
+      res.status(200).json({message: "Success!"});
     })
     .catch((err) => {
-      return throwError(err, next);
+      res.status(500).json({message: "Deleting product failed."});
     });
 };
